All formatting related to command result output are now shifted to Formatter and removed from Textui
